liquid calibration fixes

+++ src/BlankBartender.Shared/Drink.cs
@@ -7,7 +7,7 @@ namespace BlankBartender.Shared
         public byte Id { get; set; }
         public string Name { get; set; }
         public byte Type { get; set; }
-        public Dictionary<string, float> Ingradients { get; set; } = new Dictionary<string, float>();
+        public Dictionary<string, decimal> Ingradients { get; set; } = new Dictionary<string, decimal>();
         public bool IsProcessing { get; set; }
         public List<string> Garnishes { get; set; }
     }
