+++ src/BlankBartender.Shared/Pump.cs
@@ -7,7 +7,7 @@ namespace BlankBartender.Shared
         public int Number { get; set; }
         public short Pin { get; set; }
         public string Value { get; set; }
-        public int FlowRate { get; set; }
-        public float? Time { get; set; }
+        public decimal FlowRate { get; set; }
+        public decimal? Time { get; set; }
     }
 }
